requirements.txt
Flask
pymupdf
python-dotenv
pymongo
numpy
scikit-learn